+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Version 3.0.3
 -------------
 
-Unreleased
+Released 2024-04-07
 
 -   The default ``hashlib.sha1`` may not be available in FIPS builds. Don't
     access it at import time so the developer has time to change the default.